Force docker container arch
This would previously pull in aarch images when running on arm macs and we don't
have builds for linux/aarch so it would attempt to build from source.

This isn't a problem when running outside the dev container since we have
macos/aarch builds.

We can fix this on the glaredb side by just building for linux/aarch.